+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 ##### v1.1.0 LazyOption type introduction
 * Better error verbosity
 * A bit more tests
+
+##### v1.2.0 Option.ensure() static method
+* Better 'toString' representation for all options
+* More test